0.2.0 release * Add b2g support to platforms.py * We add mozci as a layer to the low-level modules * Initial documentation for mozci is added * Ignore txt, js and json files
