+++ docs/buildjson.rst
@@ -1,5 +1,5 @@
 :mod:`buildjson`
-###############
+################
 The module :mod:`buildjson`
 
 .. automodule:: mozci.buildjson
