+++ docs/mozci.rst
@@ -1,0 +1,6 @@
+:mod:`mozci`
+###############
+The module :mod:`mozci`
+
+.. automodule:: mozci.mozci
+   :members:
